LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/stretch-inline-size-003-expected.txt




















PASS .test 1
PASS .test 2
PASS .test 3
FAIL .test 4 assert_equals:
<div class="test htb" data-expected-width="55"></div>
width expected 55 but got 10
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
PASS .test 9
FAIL .test 10 assert_equals:
<div class="test htb" data-expected-width="41"></div>
width expected 41 but got 10
PASS .test 11
PASS .test 12
FAIL .test 13 assert_equals:
<div class="test htb" data-expected-width="41"></div>
width expected 41 but got 10
PASS .test 14
PASS .test 15
PASS .test 16
PASS .test 17
PASS .test 18
PASS .test 19
PASS .test 20
PASS .test 21
FAIL .test 22 assert_equals:
<div class="test htb" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
PASS .test 23
PASS .test 24
PASS .test 25
PASS .test 26
PASS .test 27
FAIL .test 28 assert_equals:
<div class="test htb" data-expected-width="41" style="box-sizing: border-box;"></div>
width expected 41 but got 10
PASS .test 29
PASS .test 30
FAIL .test 31 assert_equals:
<div class="test htb" data-expected-width="41" style="box-sizing: border-box;"></div>
width expected 41 but got 10
PASS .test 32
PASS .test 33
PASS .test 34
PASS .test 35
PASS .test 36
